feat: Build the process_data.py.
hese Python scripts should be able to run with additional arguments specifying the files used for the data and model. The script read the dataset file , clean the data, and then store it in a SQLite database.

Example:
python process_data.py disaster_messages.csv disaster_categories.csv

The Output is a db file.

Mudanças a serem submetidas:
	new file:   process_data.py