feat: Step 3. Flask Web App
We are providing much of the flask web app for you, but feel free to add extra features depending on your knowledge of flask, html, css and javascript. For this part, you'll need to:

Modify file paths for database and model as needed

The web app running with basic config.

 Mudanças a serem submetidas:
	new file:   app/run.py
	new file:   app/templates/go.html
	new file:   app/templates/master.html